NEXTEUROPA-6152: Make it possible to do a registry rebuild.
During a registry rebuild all classes are reindexed.  The language negotiation
providers are invoked during the cache clear that occurs at the end of the
rebuild, but since their paths are statically cached by drupal_get_path() at
the start of the process they are not found any more.  The easiest solution is
to simply move the language negotiation providers out of include files.